Migrate OnyxCache
